+++ cowbird/services/service.py
@@ -4,7 +4,8 @@ class Service(object):
     """
     Service interface use to notify implemented services of users/permissions changes.
 
-    TODO: Synch this interface with Magpie webhooks
+    TODO: At some point we will need a consistency function that goes through all Magpie users and make sure that
+          services are up to date.
     """
 
     def __init__(self, name, url=None):
@@ -12,7 +13,7 @@ class Service(object):
         self.url = url
 
     def json(self):
-        return {"name": self.name}
+        return {"name": self.name, "url": self.url}
 
     def create_user(self, username):
         pass
@@ -20,7 +21,7 @@ class Service(object):
     def delete_user(self, username):
         pass
 
-    def set_permission(self, permission):
+    def create_permission(self, permission):
         pass
 
     def delete_permission(self, permission):